fixed bug shoppingcart

+++ Android_Spotshop/app/src/main/java/be/fenego/android_spotshop/models/DutiesAndSurchargesTotal.java
@@ -10,7 +10,7 @@ public class DutiesAndSurchargesTotal implements Serializable {
     @SuppressWarnings("unused")
     private String type;
     @SuppressWarnings("unused")
-    private Integer value;
+    private Double value;
     @SuppressWarnings("unused")
     private String currencyMnemonic;
 
@@ -25,12 +25,12 @@ public class DutiesAndSurchargesTotal implements Serializable {
     }
 
     @SuppressWarnings("unused")
-    public Integer getValue() {
+    public double getValue() {
         return value;
     }
 
     @SuppressWarnings("unused")
-    public void setValue(Integer value) {
+    public void setValue(Double value) {
         this.value = value;
     }
